requirements.txt
fastapi
uvicorn
fyers-apiv3
python-dotenv
